rework: drop salvage scripts, demote logprob/tool-trace lenses into owning skills
- Delete the live-first-round-server.mjs one-off experiment server (unauthenticated
  spend-triggering POST /run) and its AutomationBench policy defaults.
- Strip render-logprob-compare.mjs to its generic rows.jsonl comparison mode
  (1,794 -> 471 lines): remove the hardcoded "official" experiment mode with
  baked-in run labels and literal-string call selection, the dead helper
  functions, and the duplicate CSS block. Move it under compare-trajectories.
- Demote compare-logprob-trajectories from a skill to
  compare-trajectories/references/logprob-lens.md; drop the Market Framing
  section and document the private logprob-sidecar artifact dependency.
- Demote diagnose-tool-traces from a skill to
  understand-workload/references/tool-trace-forensics.md, keeping the failure
  taxonomy and the environment-evidence vs model-visible-evidence distinction.
- Re-route the understudy orchestrator and skills/README.md to the new lenses.
- Dedupe the BF16/mlx-vlm provenance + smoke-test prose into a single canonical
  paragraph in manage-local-models/reference.md, cross-referenced from
  mlx-arena, onboard, and run-local-model-lab; break up the mlx-arena run-on
  size paragraph and normalize google/gemma-4-* checkpoint ID casing.
