+++ skills/README.md
@@ -81,7 +81,10 @@ execution; skills tell the agent what to inspect, gate, monitor, and report.
   classifies each reachable-gap task as persistence/recovery (RL-learnable),
   knowledge (not RL-addable), or format/parsing (decoding/prompt) — then counts
   the clean warm-start trajectories the comparison yields, with small-N and
-  holdout caveats.
+  holdout caveats. Its token-logprob lens
+  ([`references/logprob-lens.md`](compare-trajectories/references/logprob-lens.md))
+  covers same-row, same-family token-confidence comparison when logprob
+  sidecars exist.
 - [`pedagogical-learning`](pedagogical-learning/SKILL.md) turns privileged
   answers, execution feedback, verifier traces, or canonical solutions into
   local correct-and-learnable trajectory evidence before SFT, GRPO, or hosted RL.
@@ -89,16 +92,6 @@ execution; skills tell the agent what to inspect, gate, monitor, and report.
   stateful workload into an RLM/verifiers training surface, measures on-policy
   state coverage and surprise concentration, and decides between local
   pedagogical SFT, on-policy repair, true pedagogical RL, or verifier handoff.
-- [`compare-logprob-trajectories`](compare-logprob-trajectories/SKILL.md) is the
-  token-level complement: it joins same-row eval outputs with private logprob
-  sidecars, visualizes smaller/quantized vs larger/full-precision token streams,
-  marks the likely failure span, and labels rows for route fallback, schema
-  repair, or distillation.
-- [`diagnose-tool-traces`](diagnose-tool-traces/SKILL.md) is the trace-forensics
-  worker for environment-backed tool failures: it separates debug state from
-  model-visible evidence, reconstructs reads/writes before the first mutation,
-  classifies authority, retrieval, format, ID-resolution, parser, and harness
-  failures, and recommends the cheapest fix before training.
 - [`distill-classifier`](distill-classifier/SKILL.md) replaces an expensive
   frontier model on a classification workload (binary, multi-class,
   multi-label, structured extraction) with a fine-tuned open-weight student:
