+++ skills/understand-workload/SKILL.md
@@ -135,6 +135,16 @@ user; artifact paths for the local workload brief; and the next evidence action.
 If you include arena questions, mark them optional and tie each to a real step
 or criterion. Keep the decomposition doc local; only synthetic questions leave.
 
+## Single-run tool-failure forensics
+
+When one environment-backed or tool-calling run failed and the question is
+"why did the model make that choice from what it actually saw?", use
+[`references/tool-trace-forensics.md`](references/tool-trace-forensics.md). It
+separates environment evidence from model-visible evidence, reconstructs
+reads/writes before the first mutation, classifies the failure (retrieval,
+authority precedence, format, ID resolution, parser, harness), and recommends
+the cheapest fix before training.
+
 ## References
 
 - [`../profile-captures/SKILL.md`](../profile-captures/SKILL.md) — the fleet sibling: profile a whole capture dir into a cost + call-type taxonomy first, then aim this skill at the cluster worth decomposing.
